Fix minimap chars-per-dot ignored in pixel mode, and accept fractional values
EnsurePlane() hardcoded charsPerDot=1 for real-pixel-mode rendering
regardless of ned/set-minimap-chars-per-dot's configured value -- the
setting only ever affected the glyph-mode fallback, so anyone actually
running in pixel mode (the common case on a capable terminal) saw zero
visual effect from changing it. Both render paths now uniformly honor
editor::MinimapCharsPerDot().

Also switches the setting from int to double throughout (storage, Janet
binding, the two arithmetic spots in ForEachDensityDot) so a value between
whole-number steps (e.g. 8.5) is reachable, not just coarser integer jumps.

+++ Source/Editor/MinimapSettings.h
@@ -33,9 +33,13 @@ void              SetMinimapWidth(int columns);
 // reading as visually "overly filled" rather than showing genuine
 // per-line shape; 8 leaves the right side of the bar mostly empty for
 // typical line lengths, closer to how a real minimap should read at a
-// glance. Non-positive values clamped to 1.
-void              SetMinimapCharsPerDot(int columns);
-[[nodiscard]] int MinimapCharsPerDot();
+// glance. A floating-point value (minimap-fractional-chars-per-dot
+// follow-up) so a value between two whole-number steps (e.g. 8.5) is
+// reachable too, not just the coarser integer jumps -- real-pixel mode in
+// particular has enough columns of headroom for the difference to actually
+// show. Values <= 0 clamped to 1.
+void                 SetMinimapCharsPerDot(double columns);
+[[nodiscard]] double MinimapCharsPerDot();
 
 } // namespace ned::editor
